Corrige trips

+++ resources/views/layouts/HeaderAuth.blade.php
@@ -93,7 +93,7 @@
 
             <div class="menu _header-menu">
 
-                @if ( (int) Auth::user()->rol_id === 1 )
+                @if ( Auth::user()->rol_id === 1 )
                     <a href="{{ route('/') }}" class="menu_items hideblock743">
                         <i class="fas fa-hotel _i-red22"></i>
                         Switch to User
